requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.2
crispy-bootstrap5==2025.4
cryptography==44.0.2
cycler==0.12.1
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.16.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.57.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
passlib==1.7.4
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2

